+++ packages/mobile/src/cards/CardGroup.tsx
@@ -6,10 +6,22 @@ import { Divider } from '../layout/Divider';
 import type { GroupProps, RenderGroupItem } from '../layout/Group';
 import { Group } from '../layout/Group';
 
+/**
+ * @deprecated Use `Box`, `HStack` or `VStack` instead. This will be removed in a future major release.
+ * @deprecationExpectedRemoval v10
+ */
 export type CardGroupBaseProps = GroupProps;
+/**
+ * @deprecated Use `Box`, `HStack` or `VStack` instead. This will be removed in a future major release.
+ * @deprecationExpectedRemoval v10
+ */
 export type CardGroupProps = CardGroupBaseProps;
 export type CardGroupRenderItem = RenderGroupItem;
 
+/**
+ * @deprecated Use `Box`, `HStack` or `VStack` instead. This will be removed in a future major release.
+ * @deprecationExpectedRemoval v10
+ */
 export const CardGroup = memo(
   forwardRef<View, CardGroupProps>(function CardGroup(
     {