+++ packages/mobile/src/buttons/Button.tsx
@@ -127,8 +127,16 @@ export const Button = memo(
       borderColor,
       borderWidth = 100,
       borderRadius = compact ? 700 : 900,
+      height = interactableHeight[compact ? 'compact' : 'regular'],
       accessibilityLabel,
       accessibilityHint,
+      padding,
+      paddingStart,
+      paddingEnd,
+      paddingTop,
+      paddingBottom,
+      paddingX: paddingXProp,
+      paddingY: paddingYProp,
       ...props
     }: ButtonProps,
     ref: React.ForwardedRef<View>,
@@ -147,8 +155,6 @@ export const Button = memo(
 
     const sizingStyle = block ? styles.block : styles.inline;
     const justifyContent = flush ? 'flex-start' : hasIcon ? 'space-between' : 'center';
-
-    const minHeight = interactableHeight[compact ? 'compact' : 'regular'];
 
     const { paddingX, paddingY, marginStart, marginEnd } = getButtonSpacingProps({
       compact,
@@ -199,6 +205,7 @@ export const Button = memo(
         borderRadius={borderRadius}
         borderWidth={borderWidth}
         feedback={feedback}
+        height={height}
         loading={loading}
         marginEnd={marginEnd}
         marginStart={marginStart}
@@ -212,9 +219,14 @@ export const Button = memo(
           alignItems="center"
           flexWrap="nowrap"
           justifyContent={justifyContent}
-          minHeight={minHeight}
-          paddingX={paddingX}
-          paddingY={paddingY}
+          minHeight={height}
+          padding={padding}
+          paddingBottom={paddingBottom}
+          paddingEnd={paddingEnd}
+          paddingStart={paddingStart}
+          paddingTop={paddingTop}
+          paddingX={paddingXProp ?? paddingX}
+          paddingY={paddingYProp ?? paddingY}
           style={sizingStyle}
         >
           {loading ? (